synapse-report-generator/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
google-generativeai==0.3.2
jinja2==3.1.3
pydantic==2.6.0
